Food Thrift Android Application
There is an estimate that 70billion pounds of food is being wasted in
all the phases such as
production, transportation and consumption every year. Our application
is an idea to reduce
food wastage in the consumption phase, especially waste due to perceived
quality of food
approaching “sell by” or “use by” label. Food sellers add the data of
the food and its reduced
selling price that are about to near the sell by or use by date into the
application. The consumers
bookmark a food seller and food items they are interested in. Whenever
food seller adds data
into the application, the users get notified of the new addition if the
seller and the food item is
in their bookmarks. Hence the interested users buy the food at reduced
price and reduce the
food wastage at the same time.